Reorganise how interfaces are used
